% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_bar.R
\name{quick_bar}
\alias{quick_bar}
\title{Create a quick barplot}
\usage{
quick_bar(data, x)
}
\arguments{
\item{data}{A data frame.}

\item{x}{A categorical variable.}
}
\value{
A ggplot bar chart.
}
\description{
Create a quick barplot
}
\examples{
quick_bar(mtcars, cyl)
}
