% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_boxplot.R
\name{quick_boxplot}
\alias{quick_boxplot}
\title{Create a quick boxplot}
\usage{
quick_boxplot(data, x, y)
}
\arguments{
\item{data}{A data frame.}

\item{x}{A grouping variable.}

\item{y}{A numeric variable.}
}
\value{
A ggplot boxplot.
}
\description{
Create a quick boxplot
}
\examples{
quick_boxplot(mtcars, cyl, mpg)
}
