% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_hist.R
\name{quick_hist}
\alias{quick_hist}
\title{Create a quick histogram}
\usage{
quick_hist(data, x)
}
\arguments{
\item{data}{A data frame.}

\item{x}{A numeric variable.}
}
\value{
A ggplot histogram.
}
\description{
Create a quick histogram
}
\examples{
quick_hist(mtcars, mpg)
}
