% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_scatter.R
\name{quick_scatter}
\alias{quick_scatter}
\title{Create a quick scatterplot}
\usage{
quick_scatter(data, x, y)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Variable for the x-axis.}

\item{y}{Variable for the y-axis.}
}
\value{
A ggplot scatterplot.
}
\description{
Create a quick scatterplot
}
\examples{
quick_scatter(mtcars, wt, mpg)
}
