% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jviz.R
\name{theme_jviz}
\alias{theme_jviz}
\title{Clean theme to jviz}
\usage{
theme_jviz()
}
\value{
A ggplot2 theme.
}
\description{
Clean theme to jviz
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) +
  ggplot2::geom_point() +
  theme_jviz()
}
